+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+- Add `PG_SSL_MODE` environment variable to allow SSL connections between the API and database.
 
 ### Changed
 - Values for "Height", "Angle", "Pitch", and coordinates are now input fields. Users can adjust values using arrow keys.